requirements.txt
Django==2.2.12
django-ckeditor==5.9.0
django-filter==2.2.0
django-js-asset==1.2.2
djangorestframework==3.11.0
Markdown==3.2.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
